requirements.txt
Django
gunicorn
Pillow
psycopg2-binary
dj-database-url
python-dotenv
whitenoise